+++ src/xr_3da/x_ray.cpp
@@ -1291,7 +1291,7 @@ u32 calc_progress_color(u32 idx, u32 total, int stage, int max_stage)
 
 
 	float kk			= (float(stage+1)/float(max_stage))*(total/2.0f);
-	float f				= 1/(exp((float(idx)-kk)*0.5f)+1.0f);
+	float f				= 1/((float) exp((float(idx)-kk)*0.5f)+1.0f);
 
 	return color_argb_f		(f,1.0f,1.0f,1.0f);
 }
